Added new profile for PoE-AP
Added menuconfig for TX/RX pin config of C6 firmware.
Added free PSRAM status on website.
Fixed SDCard size display on website.

+++ ESP32_AP-Flasher/include/storage.h
@@ -27,7 +27,7 @@ class DynStorage {
     void begin();
     void end();
     void listFiles();
-    size_t freeSpace();
+    uint64_t freeSpace();
 
    private:
     bool isInited;
@@ -39,5 +39,3 @@ extern fs::FS *contentFS;
 extern void copyFile(File in, File out);
 
 #endif
-
-
